new tip added
As much as I understood the idea of merge sort from the sources, it was quite challenging to transfer it into code. I had to simplify my thinking about all of this and start with a very simple merge(array1, array2) function and then implement mergeSort. This made completing the task quicker.